move test somewhere + refine docstring
