+++ psutil/tests/test_process.py
@@ -1496,53 +1496,6 @@ class TestProcess(PsutilTestCase):
         sproc.communicate()
         self.assertEqual(sproc.returncode, 0)
 
-    def test_pid_exists_and_process_agree(self):
-        def is_linux_tid(pid):
-            try:
-                f = open("/proc/%s/status" % pid, "rb")
-            except FileNotFoundError:
-                return False
-            else:
-                with f:
-                    for line in f:
-                        if line.startswith(b"Tgid:"):
-                            tgid = int(line.split()[1])
-                            # If tgid and pid are different then we're
-                            # dealing with a process TID.
-                            return tgid != pid
-                    raise ValueError("'Tgid' line not found in %s" % path)
-
-        def check(pid):
-            # Try 3 times to avoid race conditions, especially in CI
-            # envs where PIDs may appear and disappear all of the
-            # sudden.
-            x = 3
-            while True:
-                exists = psutil.pid_exists(pid)
-                try:
-                    if exists:
-                        psutil.Process(pid)
-                        self.assertIn(pid, psutil.pids())
-                    else:
-                        with self.assertRaises(psutil.NoSuchProcess):
-                            psutil.Process(pid)
-                        self.assertNotIn(pid, psutil.pids())
-                except (psutil.Error, AssertionError) as err:
-                    x -= 1
-                    if x == 0:
-                        raise
-                else:
-                    return
-
-        with mock.patch("psutil._psplatform.debug"):  # quite verbose on win
-            for pid in range(1, 3000):
-                if LINUX and is_linux_tid(pid):
-                    # On Linux a TID (thread ID) can be passed to the
-                    # Process class and is querable like a PID (process
-                    # ID). Skip it.
-                    continue
-                check(pid)
-
 
 # ===================================================================
 # --- Limited user tests
